requirements.txt
Faker==8.1.4
kafka-python==2.0.2
python-dateutil==2.8.1
six==1.16.0
text-unidecode==1.3

